+++ CHANGELOG.md
@@ -14,7 +14,6 @@
 ## Improvements
 
 - PR #2103 Move old `column` and `bitmask` files into `legacy/` directory
-
 - PR #2109 added name to Python column classes
 - PR #1947 Cleanup serialization code
 - PR #2125 More aggregate in java API
@@ -42,6 +41,7 @@
 - PR #2180 Fix issue with isort reordering `importorskip` below imports depending on them
 - PR #2187 fix to honor dtype when numpy arrays are passed to columnops.as_column
 - PR #2190 Fix issue in astype conversion of string column to 'str'
+- PR #XXXX Fix issue with calling `head()` on one row dataframe
 
 
 # cuDF 0.8.0 (27 June 2019)